fixing toggle function
