+++ src/app/tests/page.js
@@ -1,9 +1,12 @@
+import SignUpComponent from "@/components/signup_comp";
+
 function TestPage() {
   return (
     <>
       <div className="flex items-center justify-center h-screen">
         <h1 className="text-6xl font-bold">Test Page {"==>"} Visit /tests</h1>
       </div>
+      <SignUpComponent />
     </>
   );
 }
